+++ packages/pied-web/plugin/pied-web-ux/index.php
@@ -4,8 +4,8 @@ class PiedWebUxPlugin extends \RainLoop\Plugins\AbstractPlugin
 {
     const NAME = 'Pied Web UX',
         AUTHOR = 'Pied Web',
-        VERSION = '1.10.3',
-        RELEASE = '2026-09-20',
+        VERSION = '1.10.4',
+        RELEASE = '2026-09-21',
         REQUIRED = '2.38.2',
         LICENSE = 'AGPL v3',
         DESCRIPTION = 'Accessible message actions for the Pied Web theme, using native mail commands.';